Flatten aggtor into just internal
